+++ app/routers/word_meaning.py
@@ -6,7 +6,7 @@ from app import models, schemas
 from typing import List
 from app.utils.converter import access_to_int
 from app.utils.lang import isDevanagariWord
-from app.middleware import auth_middleware
+from app.middleware import auth_middleware, logger_middleware
 
 
 router = APIRouter(
@@ -48,9 +48,8 @@ def get_word_meaning(word: str, meaning_id: int, db: Session = Depends(get_db)):
     return db_meaning
 
 
-
 @router.post("/{word}/meanings", status_code=status.HTTP_201_CREATED)
-def create_word_meaning(word: str, meaning: schemas.MeaningCreate, db: Session = Depends(get_db), current_user: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
+async def create_word_meaning(word: str, meaning: schemas.MeaningCreate, db: Session = Depends(get_db), current_user: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
     if access_to_int(current_user.access) < access_to_int(schemas.Access.READ_WRITE):
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
     
@@ -66,12 +65,14 @@ def create_word_meaning(word: str, meaning: schemas.MeaningCreate, db: Session =
     db.add(new_meaning)
     db.commit()
     db.refresh(new_meaning)
+
+    await logger_middleware.log_database_operations("meanings", new_meaning.id, "CREATE", current_user.email, new_meaning.meaning)
     
     return JSONResponse(status_code=status.HTTP_201_CREATED, content={"message": "Meaning created successfully"})
 
 
 @router.put("/{word}/meanings/{meaning_id}", status_code=status.HTTP_204_NO_CONTENT)
-def update_word_meaning(word: str, meaning_id: int, meaning: schemas.MeaningCreate, db: Session = Depends(get_db), current_user: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
+async def update_word_meaning(word: str, meaning_id: int, meaning: schemas.MeaningCreate, db: Session = Depends(get_db), current_user: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
     if access_to_int(current_user.access) < access_to_int(schemas.Access.READ_WRITE_MODIFY):
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
     
@@ -87,9 +88,11 @@ def update_word_meaning(word: str, meaning_id: int, meaning: schemas.MeaningCrea
     db_meaning.meaning = meaning.meaning
     db.commit()
 
+    await logger_middleware.log_database_operations("meanings", meaning_id, "UPDATE", current_user.email, meaning.meaning)
+
 
 @router.delete("/{word}/meanings/{meaning_id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_word_meaning(word: str, meaning_id: int, db: Session = Depends(get_db), current_user: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
+async def delete_word_meaning(word: str, meaning_id: int, db: Session = Depends(get_db), current_user: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
     if access_to_int(current_user.access) < access_to_int(schemas.Access.ALL):
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
     
@@ -101,12 +104,19 @@ def delete_word_meaning(word: str, meaning_id: int, db: Session = Depends(get_db
     if not db_word:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Word - {word} not found")
 
+    db_meaning = db.query(models.Meaning).filter(models.Meaning.id == meaning_id).first()
+    if not db_meaning:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Meaning - {meaning_id} not found")  
+
+    
     db.query(models.Meaning).filter(models.Meaning.id == meaning_id).delete()
     db.commit()
 
+    await logger_middleware.log_database_operations("meanings", meaning_id, "DELETE", current_user.email, db_meaning.meaning)
+
 
 @router.delete("/{word}/meanings", status_code=status.HTTP_204_NO_CONTENT)
-def delete_word_meanings(word: str, db: Session = Depends(get_db), current_user: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
+async def delete_word_meanings(word: str, db: Session = Depends(get_db), current_user: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
     if access_to_int(current_user.access) < access_to_int(schemas.Access.ALL):
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
     
@@ -120,3 +130,5 @@ def delete_word_meanings(word: str, db: Session = Depends(get_db), current_user:
 
     db.query(models.Meaning).filter(models.Meaning.sanskrit_word_id == db_word.id).delete()
     db.commit()
+
+    await logger_middleware.log_database_operations("meanings", db_word.id, "DELETE_ALL", current_user.email)
