+++ app/routers/word_nyaya_text_references.py
@@ -1,11 +1,12 @@
 from typing import List
 from fastapi import APIRouter, Depends, status, HTTPException
+from fastapi.responses import JSONResponse
 from sqlalchemy.orm import Session
 from app import models, schemas
 from app.utils.converter import access_to_int
 from app.utils.lang import isDevanagariWord
 from app.database import get_db
-from app.middleware import auth_middleware
+from app.middleware import auth_middleware, logger_middleware
 
 
 router = APIRouter(
@@ -48,7 +49,7 @@ def get_word_reference_nyaya_text(word: str, meaning_id: int, nyaya_text_id: int
 
 
 @router.post("/{word}/{meaning_id}/nyaya-text-references", status_code=status.HTTP_201_CREATED)
-def create_word_reference_nyaya_text(word: str, meaning_id: int, reference_nyaya_text: schemas.NyayaTextReference, db: Session = Depends(get_db), current_db_manager: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
+async def create_word_reference_nyaya_text(word: str, meaning_id: int, reference_nyaya_text: schemas.NyayaTextReference, db: Session = Depends(get_db), current_db_manager: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
     if access_to_int(current_db_manager.access) < access_to_int(schemas.Access.READ_WRITE):
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
     
@@ -65,9 +66,13 @@ def create_word_reference_nyaya_text(word: str, meaning_id: int, reference_nyaya
     db.commit()
     db.refresh(new_reference_nyaya_text)
 
+    await logger_middleware.log_database_operations("reference_nyaya_texts", new_reference_nyaya_text.id, "CREATE", current_db_manager.email, f"{new_reference_nyaya_text.source} - {new_reference_nyaya_text.description}")
+
+    return JSONResponse(status_code=status.HTTP_201_CREATED, content={"message": "Reference Nyaya Text created successfully"})
+
 
 @router.put("/{word}/{meaning_id}/nyaya-text-references/{nyaya_text_id}", status_code=status.HTTP_204_NO_CONTENT)
-def update_word_reference_nyaya_text(word: str, meaning_id: int, nyaya_text_id: int, reference_nyaya_text: schemas.NyayaTextReference, db: Session = Depends(get_db), current_db_manager: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
+async def update_word_reference_nyaya_text(word: str, meaning_id: int, nyaya_text_id: int, reference_nyaya_text: schemas.NyayaTextReference, db: Session = Depends(get_db), current_db_manager: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
     if access_to_int(current_db_manager.access) < access_to_int(schemas.Access.READ_WRITE_MODIFY):
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
     
@@ -89,9 +94,10 @@ def update_word_reference_nyaya_text(word: str, meaning_id: int, nyaya_text_id: 
     
     db.commit()
 
+    await logger_middleware.log_database_operations("reference_nyaya_texts", nyaya_text_id, "UPDATE", current_db_manager.email, f"{reference_nyaya_text.source} - {reference_nyaya_text.description}")
 
 @router.delete("/{word}/{meaning_id}/nyaya-text-references/{nyaya_text_id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_word_reference_nyaya_text(word: str, meaning_id: int, nyaya_text_id: int, db: Session = Depends(get_db), current_db_manager: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
+async def delete_word_reference_nyaya_text(word: str, meaning_id: int, nyaya_text_id: int, db: Session = Depends(get_db), current_db_manager: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
     if access_to_int(current_db_manager.access) < access_to_int(schemas.Access.ALL):
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
     
@@ -103,12 +109,19 @@ def delete_word_reference_nyaya_text(word: str, meaning_id: int, nyaya_text_id: 
     if not db_word:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Word - {word} not found")
     
+    db_reference_nyaya_text = db.query(models.ReferenceNyayaText).filter(models.ReferenceNyayaText.meaning_id == meaning_id, models.ReferenceNyayaText.id == nyaya_text_id).first()
+
+    if not db_reference_nyaya_text:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Reference Nyaya Text - {nyaya_text_id} not found")
+    
     db.query(models.ReferenceNyayaText).filter(models.ReferenceNyayaText.meaning_id == meaning_id, models.ReferenceNyayaText.id == nyaya_text_id).delete()
     db.commit()
 
+    await logger_middleware.log_database_operations("reference_nyaya_texts", nyaya_text_id, "DELETE", current_db_manager.email, f"{db_reference_nyaya_text.source}")
+
 
 @router.delete("/{word}/{meaning_id}/nyaya-text-references", status_code=status.HTTP_204_NO_CONTENT)
-def delete_word_nyaya_text_references(word: str, meaning_id: int, db: Session = Depends(get_db), current_db_manager: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
+async def delete_word_nyaya_text_references(word: str, meaning_id: int, db: Session = Depends(get_db), current_db_manager: schemas.DBManager = Depends(auth_middleware.get_current_db_manager)):
     if access_to_int(current_db_manager.access) < access_to_int(schemas.Access.ALL):
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
     
@@ -123,3 +136,5 @@ def delete_word_nyaya_text_references(word: str, meaning_id: int, db: Session = 
     
     db.query(models.ReferenceNyayaText).filter(models.ReferenceNyayaText.sanskrit_word_id == db_word.id, models.ReferenceNyayaText.meaning_id == meaning_id).delete()
     db.commit()
+
+    await logger_middleware.log_database_operations("reference_nyaya_texts", meaning_id, "DELETE", current_db_manager.email, "ALL")
